shareページのハンドラーを別ファイルに分割
