+++ route/app.ts
@@ -2,12 +2,10 @@ import { Hono } from "hono";
 import { languageDetector } from "hono/language";
 import apiApp from "./api/app.js";
 import { Bindings } from "./env.js";
-import briefApp from "./api/brief.js";
-import { ChartBrief } from "../chartFormat/chart.js";
 import { fetchStatic } from "./static.js";
-import { getTranslations } from "../i18n/i18n.js";
 import { HTTPException } from "hono/http-exception";
 import ogApp from "./og.js";
+import shareHandler from "./share.js";
 
 async function errorResponse(
   origin: string,
@@ -85,79 +83,24 @@ const app = new Hono<{ Bindings: Bindings }>({ strict: false })
       307
     );
   })
-  .on(
-    "get",
-    [
-      "/share/:cid{[0-9]+}",
-      // "/share/:cid_txt{[0-9]+\\.txt}",
-      "_next/static/chunks/app/:locale/share/:cid{[0-9]+}/:f",
-    ],
-    async (c) => {
-      const lang = c.get("language");
-      const cid = c.req.param("cid"); /*|| c.req.param("cid_txt").slice(0, -4)*/
-      const pBriefRes = briefApp.request(`/${cid}`);
-      const t = await getTranslations(lang, "share");
-      let placeholderUrl: URL;
-      if (c.req.path.startsWith("/share")) {
-        placeholderUrl = new URL(
-          `/${lang}/share/placeholder`,
-          new URL(c.req.url).origin
-        );
-      } else {
-        placeholderUrl = new URL(
-          c.req.url.replace(/share\/[0-9]+/, "share/placeholder")
-        );
-      }
-      const pRes = fetchStatic(placeholderUrl);
-      const briefRes = await pBriefRes;
-      if (briefRes.ok) {
-        const brief = (await briefRes.json()) as ChartBrief;
-        const res = await pRes;
-        const newTitle = brief.composer
-          ? t("titleWithComposer", {
-              title: brief.title,
-              composer: brief.composer,
-              chartCreator: brief.chartCreator,
-              cid: cid,
-            })
-          : t("title", {
-              title: brief.title,
-              chartCreator: brief.chartCreator,
-              cid: cid,
-            });
-        let titleEscapedJsStr = ""; // "{...\"TITLE\"}" inside script tag
-        let titleEscapedHtml = ""; // <title>TITLE</title>, "TITLE" inside meta tag
-        for (let i = 0; i < newTitle.length; i++) {
-          titleEscapedJsStr +=
-            "\\\\u" + newTitle.charCodeAt(i).toString(16).padStart(4, "0");
-          titleEscapedHtml += "&#" + newTitle.charCodeAt(i) + ";";
-        }
-        const replacedBody = (await res.text())
-          .replaceAll("/share/placeholder", `/share/${cid}`)
-          .replaceAll('\\"PLACEHOLDER_TITLE', '\\"' + titleEscapedJsStr)
-          .replaceAll("PLACEHOLDER_TITLE", titleEscapedHtml)
-          .replaceAll(
-            // これはjsファイルの中にしか現れないのでエスケープの必要はない
-            '"PLACEHOLDER_BRIEF"',
-            JSON.stringify(JSON.stringify(brief))
-          );
-        return c.text(replacedBody, 200, {
-          "Content-Type": res.headers.get("Content-Type") || "text/plain",
-          "Cache-Control": "no-store",
-        });
-      } else {
-        let message = "";
-        try {
-          message =
-            ((await briefRes.json()) as { message?: string }).message || "";
-        } catch {
-          //
-        }
-        throw new HTTPException(briefRes.status as 401 | 404 | 500, {
-          message,
-        });
-      }
+  .get("/share/:cid{[0-9]+}", async (c, next) => {
+    // @ts-expect-error TODO how to specify Variable type only for this middleware?
+    c.set("cid", c.req.param("cid"));
+    // c.req.param("cid_txt").slice(0, -4) for /share/:cid_txt{[0-9]+.txt}
+    await next();
+  })
+  .get(
+    "/_next/static/chunks/app/:locale/share/:cid{[0-9]+}/:f",
+    async (c, next) => {
+      // @ts-expect-error TODO same as above
+      c.set("cid", c.req.param("cid"));
+      await next();
     }
+  )
+  .get("/share/:cid{[0-9]+}", ...shareHandler)
+  .get(
+    "/_next/static/chunks/app/:locale/share/:cid{[0-9]+}/:f",
+    ...shareHandler
   );
 
 export default app;
